add ProducersManager, add Value object for communicating between view and model

+++ src/main/java/factory/workers/DelaysWorker.java
@@ -9,7 +9,7 @@ abstract public class DelaysWorker implements Runnable, ValueCommandRecipient {
         this.waitTime = waitTime;
     }
 
-    public double getWaitTime() {
+    public long getWaitTime() {
         return waitTime;
     }
